+++ apps/www/src/app/(main)/(docs)/components/input/examples/clearable.tsx
@@ -1,5 +1,4 @@
 import { Input } from "@jamsr-ui/react";
-import { EyeClosedIcon } from "@jamsr-ui/shared-icons";
 
 export const InputClearable = () => {
   return (
@@ -8,7 +7,6 @@ export const InputClearable = () => {
       placeholder="Enter your username"
       type="text"
       isClearable
-      endContent={<EyeClosedIcon />}
     />
   );
 };