+++ apps/www/tailwind.config.ts
@@ -1,5 +1,7 @@
 import { type Config } from "@jamsr-ui/tailwind-config";
-import { withJamsrUI } from "@jamsr-ui/theme";
+// import { withJamsrUI } from "@jamsr-ui/theme";
+import { withJamsrUI } from "../../packages/theme/src/index";
+// import { withJamsrUI } from "../../packages/theme/src/index";
 
 const config: Config = withJamsrUI({
   darkMode: "class",