+++ packages/components/alert/src/style.ts
@@ -1,6 +1,7 @@
 import type { VariantProps } from "@jamsr-ui/utils";
 import { tv } from "@jamsr-ui/utils";
 
+export const test = "";
 export const alertVariant = tv({
   slots: {
     wrapper: "relative flex gap-2 rounded-lg p-4 text-sm",
@@ -23,44 +24,44 @@ export const alertVariant = tv({
       },
       danger: { wrapper: "border border-danger-600 text-danger-500" },
       info: { wrapper: "border border-blue-600 text-blue-500" },
-      default: { wrapper: "border border-default-600 text-default-500" },
+      default: { wrapper: "border border-divider text-default-500" },
     },
   },
   compoundVariants: [
-    // filled
+    // solid
     {
       status: "danger",
       variant: "solid",
       className: {
-        wrapper: "bg-danger-100 text-danger-600",
+        wrapper: "bg-danger/10 text-danger-600",
       },
     },
     {
       status: "success",
       variant: "solid",
       className: {
-        wrapper: "bg-success-100 text-success-600",
+        wrapper: "bg-success/10 text-success-600",
       },
     },
     {
       status: "info",
       variant: "solid",
       className: {
-        wrapper: "bg-blue-100 text-blue-600",
+        wrapper: "bg-blue-600/10 text-blue-600",
       },
     },
     {
       status: "warning",
       variant: "solid",
       className: {
-        wrapper: "bg-warning-100 text-warning-600",
+        wrapper: "bg-warning/10 text-warning-600",
       },
     },
     {
       status: "default",
       variant: "solid",
       className: {
-        wrapper: "border-divider bg-default-100 text-default-600",
+        wrapper: "bg-default/10 text-default-600",
       },
     },
   ],