+++ packages/components/checkbox/src/style.ts
@@ -9,8 +9,9 @@ export const checkbox = tv({
       "ui-readonly:pointer-events-none",
     ],
     wrapper: "flex items-center gap-2",
-    label: "text-foreground shrink-0 select-none text-sm font-normal",
-    helperText: "text-foreground-600 text-xs",
+    label:
+      "shrink-0 cursor-pointer select-none text-sm font-normal text-foreground",
+    helperText: "text-xs text-foreground-600",
     checkbox: [
       "relative size-5 cursor-pointer appearance-none rounded-md border-2 transition-all duration-500 checked:border-blue-500 checked:bg-blue-500",
     ],