+++ packages/theme/src/plugin.ts
@@ -49,11 +49,11 @@ export const jamsrUiPlugins = (config?: UIThemeConfig) => {
           },
         });
         // variants
-        addVariant("ui-disabled", "&[data-disabled=true]");
-        addVariant("ui-readonly", "&[data-readonly=true]");
+        addVariant("ui-group-hover", ":merge(.group)[data-hovered=true] &");
         addVariant("ui-hover", "&[data-hovered=true]");
-        addVariant("ui-pressed", "&[data-pressed=true]");
+        addVariant("ui-group-focus", ":merge(.group)[data-focused=true] &");
         addVariant("ui-focus", "&[data-focused=true]");
+        addVariant("ui-pressed", "&[data-pressed=true]");
         addVariant("ui-active", "&[data-active=true]");
         addVariant("ui-open", "&[data-open=true]");
         addVariant("ui-checked", "&[data-checked=true]");
@@ -62,9 +62,11 @@ export const jamsrUiPlugins = (config?: UIThemeConfig) => {
         addVariant("ui-selected", "&[data-selected=true]");
         addVariant("ui-filled-within", "&[data-filled-within=true]");
         addVariant("ui-focus-visible", "&[data-focus-visible=true]");
+        addVariant("ui-readonly", "&[data-readonly=true]");
+        addVariant("ui-disabled", "&[data-disabled=true]");
         addUtilities({
-          ".text-transform-inherit":{
-            "text-transform": "inherit"
+          ".text-transform-inherit": {
+            "text-transform": "inherit",
           },
           ".tap-highlight-transparent": {
             "-webkit-tap-highlight-color": "transparent",