+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @jamsr-ui/dialog
 
+## 0.0.81
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/button@0.0.72
+  - @jamsr-ui/motion@0.0.2
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/hooks@0.0.45
+  - @jamsr-ui/shared-icons@0.0.59
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.80
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.1
+  - @jamsr-ui/utils@0.0.63
+  - @jamsr-ui/button@0.0.71
+
 ## 0.0.79
 
 ### Patch Changes