+++ packages/components/divider/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jamsr-ui/divider
 
+## 0.0.70
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.69
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.68
 
 ### Patch Changes