+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @jamsr-ui/input
 
+## 0.0.92
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/button@0.0.72
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/hooks@0.0.45
+  - @jamsr-ui/shared-icons@0.0.59
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.91
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.63
+  - @jamsr-ui/button@0.0.71
+
 ## 0.0.90
 
 ### Patch Changes