+++ packages/components/kbd/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jamsr-ui/kbd
 
+## 0.0.5
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.3
 
 ### Patch Changes