+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @jamsr-ui/menu
 
+## 0.0.88
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.2
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/hooks@0.0.45
+  - @jamsr-ui/shared-icons@0.0.59
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.87
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.1
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.86
 
 ### Patch Changes