+++ packages/components/progress/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @jamsr-ui/progress
 
+## 0.0.67
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.2
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.66
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.1
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.65
 
 ### Patch Changes