+++ packages/components/repeater/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/repeater
 
+## 0.0.37
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.36
 
 ### Patch Changes