+++ packages/components/show/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/show
 
+## 0.0.2
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.1
 
 ### Patch Changes