+++ packages/components/sidebar/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @jamsr-ui/sidebar
 
+## 0.0.77
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/shared-icons@0.0.59
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.76
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.75
 
 ### Patch Changes