+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @jamsr-ui/switch
 
+## 0.0.77
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/typography@0.0.75
+  - @jamsr-ui/motion@0.0.2
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/hooks@0.0.45
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.76
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.1
+  - @jamsr-ui/utils@0.0.63
+  - @jamsr-ui/typography@0.0.74
+
 ## 0.0.75
 
 ### Patch Changes