+++ packages/components/tab/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @jamsr-ui/tab
 
+## 0.0.89
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.2
+  - @jamsr-ui/card@0.0.80
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/hooks@0.0.45
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.88
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/motion@0.0.1
+  - @jamsr-ui/utils@0.0.63
+  - @jamsr-ui/card@0.0.79
+
 ## 0.0.87
 
 ### Patch Changes