+++ packages/components/table/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jamsr-ui/table
 
+## 0.0.49
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.63
+
 ## 0.0.47
 
 ### Patch Changes