+++ packages/config/prettier-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/prettier-config
 
+## 0.0.21
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.20
 
 ### Patch Changes