+++ packages/config/tailwind-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/tailwind-config
 
+## 0.0.22
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.21
 
 ### Patch Changes