+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/hooks
 
+## 0.0.45
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.44
 
 ### Patch Changes