+++ packages/react/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @jamsr-ui/react
 
+## 0.0.191
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/file-upload-single@0.0.91
+  - @jamsr-ui/file-upload-multi@0.0.91
+  - @jamsr-ui/autocomplete@0.0.98
+  - @jamsr-ui/confirmation@0.0.61
+  - @jamsr-ui/data-table@0.0.56
+  - @jamsr-ui/tags-input@0.0.28
+  - @jamsr-ui/typography@0.0.75
+  - @jamsr-ui/accordion@0.0.78
+  - @jamsr-ui/otp-input@0.0.68
+  - @jamsr-ui/checkbox@0.0.68
+  - @jamsr-ui/progress@0.0.67
+  - @jamsr-ui/skeleton@0.0.68
+  - @jamsr-ui/textarea@0.0.92
+  - @jamsr-ui/divider@0.0.70
+  - @jamsr-ui/popover@0.0.70
+  - @jamsr-ui/stepper@0.0.6
+  - @jamsr-ui/tooltip@0.0.67
+  - @jamsr-ui/avatar@0.0.70
+  - @jamsr-ui/button@0.0.72
+  - @jamsr-ui/dialog@0.0.81
+  - @jamsr-ui/drawer@1.0.67
+  - @jamsr-ui/editor@0.0.96
+  - @jamsr-ui/header@0.0.76
+  - @jamsr-ui/ripple@0.0.70
+  - @jamsr-ui/select@0.0.87
+  - @jamsr-ui/switch@0.0.77
+  - @jamsr-ui/alert@0.0.85
+  - @jamsr-ui/input@0.0.92
+  - @jamsr-ui/radio@0.0.76
+  - @jamsr-ui/table@0.0.49
+  - @jamsr-ui/toast@0.0.30
+  - @jamsr-ui/card@0.0.80
+  - @jamsr-ui/chip@0.0.78
+  - @jamsr-ui/link@0.0.70
+  - @jamsr-ui/menu@0.0.88
+  - @jamsr-ui/kbd@0.0.5
+  - @jamsr-ui/rhf@0.0.36
+  - @jamsr-ui/tab@0.0.89
+  - @jamsr-ui/theme@0.0.68
+  - @jamsr-ui/utils@0.0.64
+  - @jamsr-ui/core@0.0.39
+  - @jamsr-ui/dnd@0.0.10
+  - @jamsr-ui/rating@0.0.28
+  - @jamsr-ui/repeater@0.0.37
+  - @jamsr-ui/show@0.0.2
+  - @jamsr-ui/styles@0.0.9
+
+## 0.0.190
+
+### Patch Changes
+
+- fix: components
+- Updated dependencies
+  - @jamsr-ui/confirmation@0.0.60
+  - @jamsr-ui/data-table@0.0.55
+  - @jamsr-ui/accordion@0.0.77
+  - @jamsr-ui/otp-input@0.0.67
+  - @jamsr-ui/checkbox@0.0.67
+  - @jamsr-ui/progress@0.0.66
+  - @jamsr-ui/stepper@0.0.5
+  - @jamsr-ui/dialog@0.0.80
+  - @jamsr-ui/drawer@1.0.66
+  - @jamsr-ui/header@0.0.75
+  - @jamsr-ui/select@0.0.86
+  - @jamsr-ui/switch@0.0.76
+  - @jamsr-ui/alert@0.0.84
+  - @jamsr-ui/input@0.0.91
+  - @jamsr-ui/menu@0.0.87
+  - @jamsr-ui/rhf@0.0.35
+  - @jamsr-ui/tab@0.0.88
+  - @jamsr-ui/theme@0.0.67
+  - @jamsr-ui/utils@0.0.63
+  - @jamsr-ui/core@0.0.38
+  - @jamsr-ui/button@0.0.71
+  - @jamsr-ui/file-upload-multi@0.0.90
+  - @jamsr-ui/file-upload-single@0.0.90
+  - @jamsr-ui/editor@0.0.95
+  - @jamsr-ui/autocomplete@0.0.97
+  - @jamsr-ui/tags-input@0.0.27
+  - @jamsr-ui/textarea@0.0.91
+  - @jamsr-ui/avatar@0.0.69
+  - @jamsr-ui/card@0.0.79
+  - @jamsr-ui/chip@0.0.77
+  - @jamsr-ui/divider@0.0.69
+  - @jamsr-ui/kbd@0.0.4
+  - @jamsr-ui/link@0.0.69
+  - @jamsr-ui/popover@0.0.69
+  - @jamsr-ui/radio@0.0.75
+  - @jamsr-ui/rating@0.0.27
+  - @jamsr-ui/ripple@0.0.69
+  - @jamsr-ui/skeleton@0.0.67
+  - @jamsr-ui/table@0.0.48
+  - @jamsr-ui/toast@0.0.29
+  - @jamsr-ui/tooltip@0.0.66
+  - @jamsr-ui/typography@0.0.74
+
 ## 0.0.189
 
 ### Patch Changes